Add useContactPanel tests
